Move FAQ items to their own slash command file
